requirements.txt
moviepy>=1.0.3
imageio-ffmpeg>=0.4.9
Pillow>=9.0.0
numpy>=1.23.0

